requirements.txt
langchain
langchain-community
streamlit
python-dotenv
langserve
uvicorn
fastapi
sse-starlette